refactor: share `changedLineStatistics` between the tool host and change groups
